requirements.txt
openai>=1.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.50
python-dotenv>=1.0.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0